backend/reasonmljobs/requirements.txt
Django==1.11.5
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
uWSGI==2.0.15
